Door Work
Revamped a new door frame for a better testing environment and functionality. Added two new floor extensions (most likely won't be used; but added it anyway) specifically for the new door frame.

Added the new door frame as a prefab; same goes for the left and right floor door extension.

Also, just realised as I am writing this commit that the 'extention' is spelled wrong, will need to change the names from 'extention' to 'extension'. Well done to me for not being able to spell a simple word! 2 brownie points